Successfully Organize the Error Controller
