+++ app/helpers/maintenance_helper.php
@@ -1,5 +1,5 @@
 <?php
-use App\Controllers\ErrorController;
+use App\Controllers\Errors\ErrorController;
 
 if (!function_exists('check_maintenance_mode')) {
     function check_maintenance_mode(): void
